+++ features/support/shared_steps.js
@@ -31,7 +31,7 @@ module.exports = function () {
                 var afterRequest = (err, res, body) => {
                     if (err) return cb(err);
                     if (body && body.length) {
-                        var instructions, bearings, turns, modes, times, distances, summary, intersections;
+                        var destinations, instructions, bearings, turns, modes, times, distances, summary, intersections;
 
                         var json = JSON.parse(body);
 
@@ -39,6 +39,7 @@ module.exports = function () {
 
                         if (hasRoute) {
                             instructions = this.wayList(json.routes[0]);
+                            destinations = this.destinationsList(json.routes[0]);
                             bearings = this.bearingList(json.routes[0]);
                             turns = this.turnList(json.routes[0]);
                             intersections = this.intersectionList(json.routes[0]);
@@ -122,6 +123,7 @@ module.exports = function () {
                             putValue('modes', modes);
                             putValue('times', times);
                             putValue('distances', distances);
+                            putValue('destinations', destinations);
                         }
 
                         var ok = true;