+++ include/engine/datafacade/datafacade_base.hpp
@@ -142,6 +142,8 @@ class BaseDataFacade
 
     virtual std::string GetNameForID(const unsigned name_id) const = 0;
 
+    virtual std::string GetDestinationsForID(const unsigned name_id) const = 0;
+
     virtual std::size_t GetCoreSize() const = 0;
 
     virtual std::string GetTimestamp() const = 0;