+++ include/engine/guidance/assemble_steps.hpp
@@ -92,10 +92,12 @@ std::vector<RouteStep> assembleSteps(const DataFacadeT &facade,
             {
                 BOOST_ASSERT(segment_duration >= 0);
                 const auto name = facade.GetNameForID(step_name_id);
+                const auto destinations = facade.GetDestinationsForID(step_name_id);
                 const auto distance = leg_geometry.segment_distances[segment_index];
 
                 steps.push_back(RouteStep{step_name_id,
-                                          name,
+                                          std::move(name),
+                                          destinations,
                                           NO_ROTARY_NAME,
                                           segment_duration / 10.0,
                                           distance,
@@ -142,6 +144,7 @@ std::vector<RouteStep> assembleSteps(const DataFacadeT &facade,
         BOOST_ASSERT(duration >= 0);
         steps.push_back(RouteStep{step_name_id,
                                   facade.GetNameForID(step_name_id),
+                                  facade.GetDestinationsForID(step_name_id),
                                   NO_ROTARY_NAME,
                                   duration / 10.,
                                   distance,
@@ -165,6 +168,7 @@ std::vector<RouteStep> assembleSteps(const DataFacadeT &facade,
 
         steps.push_back(RouteStep{source_node.name_id,
                                   facade.GetNameForID(source_node.name_id),
+                                  facade.GetDestinationsForID(source_node.name_id),
                                   NO_ROTARY_NAME,
                                   duration / 10.,
                                   leg_geometry.segment_distances[segment_index],
@@ -189,6 +193,7 @@ std::vector<RouteStep> assembleSteps(const DataFacadeT &facade,
     BOOST_ASSERT(!leg_geometry.locations.empty());
     steps.push_back(RouteStep{target_node.name_id,
                               facade.GetNameForID(target_node.name_id),
+                              facade.GetDestinationsForID(target_node.name_id),
                               NO_ROTARY_NAME,
                               ZERO_DURATION,
                               ZERO_DISTANCE,