+++ src/engine/api/json_factory.cpp
@@ -187,6 +187,8 @@ util::json::Object makeRouteStep(guidance::RouteStep step, util::json::Value geo
     route_step.values["distance"] = std::round(step.distance * 10) / 10.;
     route_step.values["duration"] = std::round(step.duration * 10) / 10.;
     route_step.values["name"] = std::move(step.name);
+    if (!step.destinations.empty())
+      route_step.values["destinations"] = std::move(step.destinations);
     if (!step.rotary_name.empty())
         route_step.values["rotary_name"] = std::move(step.rotary_name);